% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{dplyr_methods}
\alias{dplyr_methods}
\alias{arrange.mlvocab_vocab}
\title{Methods for \code{dplyr} predicates}
\usage{
arrange.mlvocab_vocab(.data, ...)
}
\arguments{
\item{.data}{vocab data.frame}

\item{...}{other parameters}
}
\description{
Needed to circumvent dropping attributes by dplyr/tibble functions.
}
\keyword{internal}
