% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mlvocab-package}
\alias{mlvocab}
\alias{mlvocab-package}
\title{Vocabulary utilities with support for \code{magistral}.}
\description{
Vocabulary utilities with support for \code{magistral}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vspinu/mlvocab/}
  \item Report bugs at \url{https://github.com/vspinu/mlvocab/issues}
}

}
\author{
\strong{Maintainer}: Vitalie Spinu \email{spinuvit@gmail.com}

}
\keyword{internal}
