% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{tiseq}
\alias{tiseq}
\alias{timat}
\title{Convert text to integer indexes}
\usage{
tiseq(corpus, vocab, keep_unknown = unknown_buckets > 0,
  unknown_buckets = attr(vocab, "unknown_buckets"), reverse = FALSE)

timat(corpus, vocab, maxlen = 100, pad_right = TRUE, trunc_right = TRUE,
  keep_unknown = unknown_buckets > 0, unknown_buckets = attr(vocab,
  "unknown_buckets"), reverse = FALSE)
}
\arguments{
\item{corpus}{text corpus}

\item{vocab}{data frame produced by \code{\link[=vocab]{vocab()}} or \code{\link[=vocab_update]{vocab_update()}}}

\item{keep_unknown}{logical. If \code{TRUE}, preserve unknowns in the output
sequences.}

\item{unknown_buckets}{integer. How many buckets to hash unknowns into.}

\item{reverse}{logical. Should each sequence be reversed in the final
output? Reversion happens after \code{pad_right} and \code{trunc_right} have been
applied to the original text sequence. Default \code{FALSE}.}

\item{maxlen}{integer. Maximum length of each sequence.}

\item{pad_right}{logical. Should 0-padding of shorter than \code{maxlen}
sequences happen on the right? Default \code{TRUE}.}

\item{trunc_right}{logical. Should truncation of longer than \code{maxlen}
sequences happen on the right? Default \code{TRUE}.}
}
\value{
\code{\link[=tiseq]{tiseq()}} returns a list of integer vectors, \code{\link[=timat]{timat()}} returns an
integer matrix, one row per sequence.
}
\description{
Convert text to integer indexes
}
