% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm}
\alias{dtm}
\alias{tdm}
\title{Term and co-occurrence matrices}
\usage{
dtm(corpus, vocab = NULL, unknown_buckets = attr(vocab, "unknown_buckets"),
  output = c("triplet", "column", "row", "df"))

tdm(corpus, vocab = NULL, unknown_buckets = attr(vocab, "unknown_buckets"),
  output = c("triplet", "column", "row", "df"))
}
\arguments{
\item{corpus}{a list of character vectors}

\item{vocab}{a \code{data.frame} produced by an early call to \code{\link[=vocab]{vocab()}}. When
\code{vocab} is \code{NULL} and \code{unknown_buckets} is \code{NULL} or \code{0}, the vocabulary
is first computed from corpus. When \code{unknown_buckets} > \code{0} and \code{vocab} is
\code{NULL} the result matrix will consist of buckets only.}

\item{unknown_buckets}{number of unknown buckets}

\item{output}{one of "triplet", "column", "row", "df" or an unambiguous
abbreviation thereof. First three options return the corresponding sparse
matrices from Matrix package, "df" results in a triplet \code{data.frame}.}
}
\description{
These functions compute various term-counts of a corpus with flexible output
specification.
}
