% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{vocab}
\alias{vocab}
\alias{vocab_update}
\alias{vocab_prune}
\title{Manipulate vocabularies}
\usage{
vocab(corpus, ngram = c(1, 1), chargram = FALSE)

vocab_update(vocab, corpus)

vocab_prune(vocab, max_terms = Inf, term_count_min = 1L,
  term_count_max = Inf, doc_proportion_min = 0, doc_proportion_max = 1,
  doc_count_min = 1L, doc_count_max = Inf)
}
\arguments{
\item{corpus}{list of character vectors}

\item{ngram}{a vector of the form \code{c(min_ngram, max_ngram)}.}

\item{chargram}{currently unused}

\item{vocab}{\code{data.frame} obtained from a call to \code{\link[=vocab]{vocab()}} or
\code{\link[text2vec:create_vocab]{text2vec::create_vocab()}} from \code{text2vec} package.}

\item{max_terms}{max number of terms to preserve}

\item{term_count_min}{keep terms occurring at \emph{least} this many times over all docs}

\item{term_count_max}{keep terms occurring at \emph{most} this many times over all docs}

\item{doc_count_min, doc_proportion_min}{keep terms appearing in at \emph{least} this many docs}

\item{doc_count_max, doc_proportion_max}{keep terms appearing in at \emph{most} this many docs}
}
\description{
\code{\link[=vocab]{vocab()}} creates a vocabulry from a text corpus. \code{\link[=vocab_update]{vocab_update()}} updates
an existing vocabulary. \code{\link[=vocab_prune]{vocab_prune()}} prunes vocabulary.

\code{vocab_prune} is an adaptation of \code{\link[text2vec:prune_vocabulary]{text2vec::prune_vocabulary()}}.
}
