% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{tixseq}
\alias{tixseq}
\alias{tixmat}
\title{Convert text to integer indexes}
\usage{
tixseq(corpus, vocab, keep_unknown = unknown_buckets > 0,
  unknown_buckets = 0, reverse = FALSE)

tixmat(corpus, vocab, maxlen = 100, pad_right = TRUE,
  trunc_right = TRUE, keep_unknown = unknown_buckets > 0,
  unknown_buckets = 0, reverse = FALSE)
}
\arguments{
\item{corpus}{}

\item{vocab}{}

\item{keep_unknown}{}

\item{unknown_buckets}{}

\item{maxlen}{}

\item{pad_right}{}

\item{trunc_right}{}

\item{corpus}{}

\item{vocab}{}

\item{keep_unknown}{}
}
\description{
Convert text to integer indexes
}
