% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{corpus2ixseq}
\alias{corpus2ixseq}
\alias{doc2ixseq}
\alias{corpus2ixmat}
\title{Convert text to integer indexes}
\usage{
corpus2ixseq(corpus, vocab, keep_unknown = unknown_buckets > 0,
  unknown_buckets = 0)

doc2ixseq(doc, vocab, keep_unknown = unknown_buckets > 0,
  unknown_buckets = 0)

corpus2ixmat(corpus, vocab, maxlen = 100, pad_right = TRUE,
  trunc_right = TRUE, keep_unknown = unknown_buckets > 0,
  unknown_buckets = 0)
}
\arguments{
\item{corpus}{}

\item{vocab}{}

\item{keep_unknown}{}

\item{unknown_buckets}{}

\item{doc}{}

\item{maxlen}{}

\item{pad_right}{}

\item{trunc_right}{}

\item{corpus}{}

\item{vocab}{}

\item{keep_unknown}{}
}
\description{
Convert text to integer indexes
}
