% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{vocab}
\alias{vocab}
\alias{update_vocab}
\alias{prune_vocab}
\alias{embed_vocab}
\title{Manipulate vocabularies}
\usage{
vocab(corpus, ngram = c(1, 1), ngram_sep = "_")

update_vocab(vocab, corpus)

prune_vocab(vocab, max_terms = Inf, term_count_min = 1L,
  term_count_max = Inf, doc_proportion_min = 0, doc_proportion_max = 1,
  doc_count_min = 1L, doc_count_max = Inf)

embed_vocab(vocab, embeddings, unknown_buckets = 0, max_in_bucket = 30)
}
\arguments{
\item{corpus}{list of character vectors}

\item{ngram}{a vector of the form \code{c(min_ngram, max_ngram)}.}

\item{vocab}{\code{data.frame} obtained from a call to \code{\link[=vocab]{vocab()}}.}

\item{max_terms}{max number of terms to preserve}

\item{term_count_min}{keep terms occurring at \emph{least} this many times over all docs}

\item{term_count_max}{keep terms occurring at \emph{most} this many times over all docs}

\item{doc_count_min, doc_proportion_min}{keep terms appearing in at \emph{least} this many docs}

\item{doc_count_max, doc_proportion_max}{keep terms appearing in at \emph{most} this many docs}

\item{embeddings}{embeddings matrix. The terms dimension must be named. If
both \code{\link[=colnames]{colnames()}} and \code{\link[=rownames]{rownames()}} are non-null, dimension with more
elements is considered term-dimension.}

\item{unknown_buckets}{How many unknown buckets to create along the terms of
the \code{vocab}. For each unknown bucket at least \code{max_in_bucket} embedding
vectors will be hashed into the bucket and averaged.}

\item{max_in_bucket}{At most this many embedding vectors will be averaged
into each unknown or missing bucket (see details). Lower number results
in faster processing. For large \code{unknown_buckets} this number might not
be reached due to the finiteness of the \code{embeddings} vocabulary, or even
result in \code{0} embeddings being hashed into a bucket producing \code{[0 0 ...]} embeddings for some buckets.}
}
\description{
\code{\link[=vocab]{vocab()}} creates a vocabulry from a text corpus. \code{\link[=update_vocab]{update_vocab()}} updates
an existing vocabulary. \code{\link[=prune_vocab]{prune_vocab()}} prunes vocabulary.

Similarly, when a term from the vocabulary is not present in the embedding
matrix (aka missing) \code{min_to_average} embeddings are averaged to produce the
missing embedding. Different buckets are used for "missing" and "unknown"
embeddings because \code{unknown_buckets} can be 0.
}
